% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_datadoc.R
\docType{data}
\name{area}
\alias{area}
\title{Area codes}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area codes in the column 'polygonName'
}
\usage{
data(area)
}
\description{
Definitions of area codes for Norwegian fisheries statistics
}
\examples{
data(area)

}
\keyword{datasets}
