% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProsessing.R
\name{fixInclusionProb}
\alias{fixInclusionProb}
\title{Estimates inclusion probabilities}
\usage{
fixInclusionProb(RDBESsamples)
}
\arguments{
\item{RDBESsamples}{RDBES data as exemplified in \code{\link{samples}}}
}
\value{
RDBESsamples with SAinclusionProb set
}
\description{
Estimates inclusion probabilities on SA table from catch weights and individual weights
of age determined fish.
In effect considers age sample fish as a SRSWOR selection from haul
These inclusion probabilities may not be applicable for estimation with other parameters
}
\details{
Halts with error of any SA entries are missing values for SAsampleWeightLive or SAtotalWeightLive
}
