% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICESrectangles_datadoc.R
\docType{data}
\name{ICESrectangles}
\alias{ICESrectangles}
\title{Area codes}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with ICES rectangles identified in the column 'polygonName'
}
\usage{
data(ICESrectangles)
}
\description{
Definitions of ICES rectangles
}
\examples{
data(ICESrectangles)

}
\keyword{datasets}
