% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProsessing.R
\name{imputeRatios}
\alias{imputeRatios}
\title{Impute ratios}
\usage{
imputeRatios(ratios, map)
}
\arguments{
\item{ratios}{data.frame with columns SDid (int), stratum (chr), age (int), ratio (num)}

\item{map}{list mapping a stratum to another stratum with known ratio}
}
\description{
Impute ratios to strata
}
\details{
The ratio for a stratum will be replaced by the one from a different stratum according to the list
'map'. After execution the ratio for stratum A will be \code{map[A]}, whether the ratio for A was unkwon or not

Halts with error if the ratio for \code{map[A]} is NA.
}
