% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings_datadoc.R
\docType{data}
\name{landings}
\alias{landings}
\title{Example data - Norwegian landings of Mackerell and Blue Whithing 2019 in RDBES format
Data was obtained from RDBES format as of the 2020 test data called.
The exchange format was parsed with the parser at:
https://github.com/davidcurrie2001/MI_RDBES_ExchangeFiles/blob/d38b75fe6a902892943cf01100cdbb7e9b13c21a/ReadExchangeFileExample_WKRDBEST2.R}
\format{
list of 1 data frame. Variables documented in https://github.com/ices-tools-dev/RDBES
}
\usage{
data(landings)
}
\description{
Example data - Norwegian landings of Mackerell and Blue Whithing 2019 in RDBES format
Data was obtained from RDBES format as of the 2020 test data called.
The exchange format was parsed with the parser at:
https://github.com/davidcurrie2001/MI_RDBES_ExchangeFiles/blob/d38b75fe6a902892943cf01100cdbb7e9b13c21a/ReadExchangeFileExample_WKRDBEST2.R
}
\examples{
data(landings)

}
\keyword{datasets}
