% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProsessing.R
\name{maxVarianceImputation}
\alias{maxVarianceImputation}
\title{Max variance imputation}
\usage{
maxVarianceImputation(ratio_var)
}
\arguments{
\item{ratio_var}{data.frame with columns SDid (int), stratum (chr), age (int), variance (num)}
}
\value{
'ratio_var' with ratios for unsampled strata imputed
}
\description{
Assign to unsampled strata the maximal variance among the sampled strata
}
