% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProsessing.R
\name{medianRatioImputation}
\alias{medianRatioImputation}
\title{Median ratio imputation}
\usage{
medianRatioImputation(ratio)
}
\arguments{
\item{ratio}{data.frame with columns SDid (int), stratum (chr), age (int), ratio (num)}
}
\value{
'ratio' with ratios for unsampled strata imputed
}
\description{
Assign to unsampled strata the median ratio of that age groupof other strata
}
