% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{numAtAgeBV}
\alias{numAtAgeBV}
\title{Calculate number at age}
\usage{
numAtAgeBV(BV, minAge = NULL, maxAge = NULL)
}
\arguments{
\item{BV}{BV table}

\item{minAge}{minimal age to include in estimation, define from data if NULL}

\item{maxAge}{minimal age to include in estimation, define from data if NULL}
}
\value{
data.frame with columns 'SAid', 'age' and 'count', all integers
}
\description{
Calculates number of fish at each age in a provided age range,
for each sample (SAid).
}
\details{
halts with error if fish is selected in stratified manner
or if any ages are missing.
}
