% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{numAtAgeSpeciesSelectionHT}
\alias{numAtAgeSpeciesSelectionHT}
\title{Horvitz-Thompson estimator for a sample of SA entries}
\usage{
numAtAgeSpeciesSelectionHT(SA, sampleTotals)
}
\arguments{
\item{SA}{SA table}

\item{sampleTotals}{data.frame with columns 'SSid', 'age' and 'count', all integers}
}
\value{
data.frame with columns 'SSid' (int), 'age' (int), and 'total' (num)
}
\description{
Estimates total for a given species selection
}
\details{
Haults with error if SA table has missing values for some inclusion probabilites.
Haults with error if SA table stratified entries.
Haults with error if SA table has missing values for species codes or contain more than one species code.
}
