% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officialTotalMac_datadoc.R
\docType{data}
\name{officialTotalMac}
\alias{officialTotalMac}
\title{Official estimates}
\format{
data.frame with age (age:int) and estimated landed number at age (numAtAge:num)
}
\usage{
data(officialTotalMac)
}
\description{
Official estimates of landed Mackerel from Norwegian fisheries in 2019,
as calculated with the Bayesian Framework ECA (Reca) at IMR.
}
\details{
numAtAge is obtained as the mean of the posterior distribution
}
\examples{
data(officialTotalMac)

}
\keyword{datasets}
