% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officialVarianceMac_datadoc.R
\docType{data}
\name{officialVarMac}
\alias{officialVarMac}
\title{Official estimates}
\format{
data.frame with age (age:int) and estimated variance of landed number at age (variance:num)
}
\usage{
data(officialVarMac)
}
\description{
Official estimates of landed Mackerel from Norwegian fisheries in 2019,
as calculated with the Bayesian Framework ECA (Reca) at IMR.
}
\details{
variance is obtained as the variance of the posterior distribution
}
\examples{
data(officialVarMac)

}
\keyword{datasets}
