% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ratio_coverage}
\alias{plot_ratio_coverage}
\title{Plots ratio strata coverage}
\usage{
plot_ratio_coverage(ratio_vars, landings, age, cumulativeWeightCutoff = 0.99)
}
\arguments{
\item{ratio_vars}{variance of estimate of ratio: data.frame with columns 'straum' (chr), 'age' (int), 'SDid' (int) and 'variance' (num)}

\item{landings}{CE table with the column 'stratum' added}

\item{age}{age to show variance for}

\item{cumulativeWeightCutoff}{threshold for cumulative weight. Smallest strata are excluded from plot after this threshold}
}
\description{
Plots the total landing and the standard error of the ratio of number at age to catchweight in each strata
}
