% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotNumAtAge}
\alias{plotNumAtAge}
\title{Plot number at age}
\usage{
plotNumAtAge(numAtAge)
}
\arguments{
\item{numAtAge}{data.frame with the columns 'age' (int) and 'numAtAge' (num)}
}
\description{
Plot number at age
}
