% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProsessing.R
\name{postStratifyMacPelLandings}
\alias{postStratifyMacPelLandings}
\title{Annotate post-stratification landings}
\usage{
postStratifyMacPelLandings(landings)
}
\arguments{
\item{landings}{CL table}
}
\value{
'landings' with the column 'stratum' added
}
\description{
Applies a post-stratification scheme that matches the one produced by \code{\link{postStratifyMacPelSamples}}
}
