% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProsessing.R
\name{postStratifyMacPelSamples}
\alias{postStratifyMacPelSamples}
\title{Annotate post-stratification samples}
\usage{
postStratifyMacPelSamples(samples)
}
\arguments{
\item{samples}{RDBES samples}
}
\value{
'samples' with the columns 'FOstratification' and 'FOstratumName' altered
}
\description{
Applies a post-stratification scheme that matches the one produced by \code{\link{postStratifyMacPelLandings}}
}
