% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{ratio_estimate_strata}
\alias{ratio_estimate_strata}
\title{Estimates total number at age for each strata and sampling scheme.}
\usage{
ratio_estimate_strata(ratios, landings)
}
\arguments{
\item{ratios}{data.frame with columns SDid (int), stratum (chr), age (int), ratio (num)}

\item{landings}{CL table with the column 'stratum' added}
}
\value{
data.frame with columns SDid (int), stratum (chr), age (int), numAtAge (num)
}
\description{
Estimates total number at age for each strata and sampling scheme.
}
\details{
Haults with error if 'SDid' is not a column of 'ratios'
Haults with error of 'stratum' is not a column of 'landings'
Haults with error if ratios are not provided for all strata or if ratios are provided for strata not in cencus ('landings')
Haults with error if there are more than one species code in the landings
}
