% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{ratio_variance_strata}
\alias{ratio_variance_strata}
\title{Variance for ratio estimates of totals}
\usage{
ratio_variance_strata(ratio_variances, landings)
}
\arguments{
\item{ratio_variances}{variance of ratios data.frame with columns SDid (int), stratum (chr), age (int), variance (num)}

\item{landings}{CL table with the column 'stratum' added}
}
\value{
ratio_variances with variance replaced with the variance for the strata total
}
\description{
Estimates variance for totals in strata estimated by ratio estimator.
}
\details{
Haults with error if 'SDid' is not a column of 'ratios'
Haults with error of 'stratum' is not a column of 'landings'
Haults with error if ratios are not provided for all strata or if ratios are provided for strata not in cencus ('landings')
}
