% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{ratio_variance_wo_N}
\alias{ratio_variance_wo_N}
\title{Estimate variance of ratio}
\usage{
ratio_variance_wo_N(
  sampleUnitType,
  sampleTable,
  numAtAge,
  totalWeight,
  parentIdname,
  fpc = 1
)
}
\arguments{
\item{sampleUnitType}{character identifying the kind of sample units in the sample, follows the record type sin RDBES, e.g. "FO" for an FO-table.}

\item{sampleTable}{table with information about the sample (e.g. FO-table in the RBDES data model)}

\item{numAtAge}{data.frame with columns: 'age' (int), 'total' (num, estimated total for sample unit), and a column identifying the sample unit (see details).}

\item{totalWeight}{data.frame with columns: 'total' (num, estimated total for sample unit), and a column identifying the sample unit (see details).}

\item{parentIdname}{name of column sampleUnitType that identifies the parent sample in the RDBES hierachy.}

\item{fpc}{(N-n)/N, where N is population size and n is sample size. Defaults to 1, which is applicable when N >> n}
}
\value{
data.frame with columns parentname (int), stratum (chr), age (int), and variance (num)
}
\description{
Approximate estimation for the variance of the ratio between numberAtAge and weight
Adapted from Hankin, Mohr and Newman, Eq. 7.19

The 'sampleTable' is a table representing one of the leves in the RDBES hierarchy and has necessary information about stratification etc.
The 'numAtAge' table and the 'totalWeight' table contains estimate of the number at age and the catch weight for each of the unit in the sample.
In addition they contain a column identifying the sample unit. E.g 'FOid' if 'sampleTable' is an FO-table, or 'LEid' if the sampleTable is an 'LE'-table

Halts with error if the sample is a clustered sample
}
