% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{ratio_wo_N}
\alias{ratio_wo_N}
\title{Estimates ratio of number at age and weight}
\usage{
ratio_wo_N(sampleUnitType, sampleTable, numAtAge, totalWeight, parentIdname)
}
\arguments{
\item{sampleUnitType}{character identifying the kind of sample units in the sample, follows the record type sin RDBES, e.g. "FO" for an FO-table.}

\item{sampleTable}{table with information about the sample (e.g. FO-table in the RBDES data model)}

\item{numAtAge}{data.frame with columns: 'age' (int), 'total' (num, estimated total for sample unit), and a column identifying the sample unit (see details).}

\item{totalWeight}{data.frame with columns: 'total' (num, estimated total for sample unit), and a column identifying the sample unit (see details).}

\item{parentIdname}{name of column sampleUnitType that identifies the parent sample in the RDBES hierachy.}
}
\value{
data.frame with columns parentIdname (int), stratum (chr), age (int), ratio (num)
}
\description{
Estimates ratio of number at age and weight
when samples are selected SRSWOR and population totals are not available.
}
\details{
The function allows estimation from a sample of different kinds of sampling units.
The 'sampleTable' is a table representing one of the leves in the RDBES hierarchy and has necessary information about stratification etc.
The 'numAtAge' table and the 'totalWeight' table contains estimate of the number at age and the catch weight for each of the unit in the sample.
In addition they contain a column identifying the sample unit. E.g 'FOid' if 'sampleTable' is an FO-table, or 'LEid' if the sampleTable is an 'LE'-table

Halts with error if the sample is a clustered sample
}
