% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_datadoc.R
\docType{data}
\name{samples}
\alias{samples}
\title{Samples example}
\format{
list of 8 data frames. Variables documented in https://github.com/ices-tools-dev/RDBES
}
\usage{
data(samples)
}
\description{
Example data - Norwegian catch of mac.27.nea sampled in 2019 in RDBES format
Data was obtained from RDBES format as of the 2020 test data called.
The exchange format was parsed with the parser at:
https://github.com/davidcurrie2001/MI_RDBES_ExchangeFiles/blob/d38b75fe6a902892943cf01100cdbb7e9b13c21a/ReadExchangeFileExample_WKRDBEST2.R
and some fields like FO$SDid has been corrected afterwards.
}
\details{
The parameter Age on the BV table is incorrectly labelled with selection methods CENSUS for all entries, This should be SRSWOR.
}
\examples{
data(samples)

}
\keyword{datasets}
