% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{total_stratified}
\alias{total_stratified}
\title{Estimates grand total}
\usage{
total_stratified(strata_totals)
}
\arguments{
\item{strata_totals}{data.frame with columns SDid (int), stratum (chr), age (int), numAtAge (num)}
}
\value{
data.frame with columns age (int) and numAtAge (num)
}
\description{
Estimates grand total number at age from strata totals
}
