% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{totalWeightHaulHT}
\alias{totalWeightHaulHT}
\title{Calculate haul weight from species-selection totals}
\usage{
totalWeightHaulHT(SS, SStotals)
}
\arguments{
\item{SS}{SS table}

\item{SStotals}{total weight for each sample (SA)}
}
\value{
data.frame with columns 'FOid' (int), 'age' (int), and 'total' (num)
}
\description{
Calculate haul weihgt from species-selection totals.
Only support census-application of species lists
}
\details{
Haults with error if any species lists are sampled from a set of species lists
}
