% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{totalWeightSpeciesSelection}
\alias{totalWeightSpeciesSelection}
\title{Calculate total weight for species selection from SA entires}
\usage{
totalWeightSpeciesSelection(SA)
}
\arguments{
\item{SA}{SA table}
}
\value{
data.frame with columns 'SSid' (int), and 'weight' (num)
}
\description{
Calculates total for a given species selection
}
\details{
Haults with error if SA table stratified entries.
Haults with error if SA table has missing values for species codes or contain more than one species code.
}
