% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{variance_stratified}
\alias{variance_stratified}
\title{Estimates variance for grand total}
\usage{
variance_stratified(strata_variances)
}
\arguments{
\item{strata_variances}{data.frame with columns SDid (int), stratum (chr), age (int), variance (num)}
}
\value{
data.frame with columns age (int) and variance (num)
}
\description{
Estimates variance for grand total number at age from strata variances
}
