% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProsessing.R
\name{fixInclusionProb}
\alias{fixInclusionProb}
\title{Estimates inclusion probabilities}
\usage{
fixInclusionProb(samples)
}
\arguments{
\item{samples}{RDBES data as exemplified in \code{\link{samples}}}
}
\value{
samples with SAinclusionProb set
}
\description{
Estimates inclusion probabilities on SA table from catch weights and sample weights
}
\details{
Halts with error of any SA entries are missing values for SAsampleWeightLive or SAtotalWeightLive
}
