% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotNumAtAge}
\alias{plotNumAtAge}
\title{Plot number at age}
\usage{
plotNumAtAge(numAtAge, variance = NULL)
}
\arguments{
\item{numAtAge}{data.frame with the columns 'age' (int) and 'numAtAge' (num)}

\item{variance}{data.frame with columns age (int) and variance (num)}
}
\description{
Plot number at age
}
\examples{
 data(officialTotalMac)
 data(officialVarMac)
 plotNumAtAge(officialTotalMac, officialVarMac)
}
