% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{ratio_estimate_strata}
\alias{ratio_estimate_strata}
\title{Estimates total number at age for each strata and sampling scheme.}
\usage{
ratio_estimate_strata(ratios, landings)
}
\arguments{
\item{ratios}{data.frame with columns SDid (int), stratum (chr), age (int), ratio (num)}

\item{landings}{CL table with the column 'stratum' added}
}
\value{
data.frame with columns SDid (int), stratum (chr), age (int), numAtAge (num)
}
\description{
Estimates total number at age for each strata and sampling scheme.
}
